+++ packages/eslint-plugin/tests/rules/no-import-type-side-effects.test.ts
@@ -14,7 +14,6 @@ ruleTester.run('no-import-type-side-effects', rule, {
     "import { type T, U } from 'mod';",
     "import { T, type U } from 'mod';",
     "import type T from 'mod';",
-    "import type T, { U } from 'mod';",
     "import T, { type U } from 'mod';",
     "import type * as T from 'mod';",
     "import 'mod';",