+++ packages/project-service/CHANGELOG.md
@@ -1,3 +1,39 @@
+## 8.53.0 (2026-01-12)
+
+### 🚀 Features
+
+- **project-service:** allow passing `Partial<ts.server.ServerHost>` to project service ([#11932](https://github.com/typescript-eslint/typescript-eslint/pull/11932))
+
+### ❤️ Thank You
+
+- auvred @auvred
+
+You can read about our [versioning strategy](https://typescript-eslint.io/users/versioning) and [releases](https://typescript-eslint.io/users/releases) on our website.
+
+## 8.52.0 (2026-01-05)
+
+This was a version bump only for project-service to align it with other projects, there were no code changes.
+
+You can read about our [versioning strategy](https://typescript-eslint.io/users/versioning) and [releases](https://typescript-eslint.io/users/releases) on our website.
+
+## 8.51.0 (2025-12-29)
+
+This was a version bump only for project-service to align it with other projects, there were no code changes.
+
+You can read about our [versioning strategy](https://typescript-eslint.io/users/versioning) and [releases](https://typescript-eslint.io/users/releases) on our website.
+
+## 8.50.1 (2025-12-22)
+
+This was a version bump only for project-service to align it with other projects, there were no code changes.
+
+You can read about our [versioning strategy](https://typescript-eslint.io/users/versioning) and [releases](https://typescript-eslint.io/users/releases) on our website.
+
+## 8.50.0 (2025-12-15)
+
+This was a version bump only for project-service to align it with other projects, there were no code changes.
+
+You can read about our [versioning strategy](https://typescript-eslint.io/users/versioning) and [releases](https://typescript-eslint.io/users/releases) on our website.
+
 ## 8.49.0 (2025-12-08)
 
 This was a version bump only for project-service to align it with other projects, there were no code changes.